+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Latest
+
+- Add support for automation level to event data model.
+
 ## 1.0.3
 
 - Update Uri parsing for newer versions of dart.